Annotate masses

This function annotates a feature table based on exact mass match. It requires a structural library, its metadata, and lists of adducts, clusters, and neutral losses to be considered. The polarity has to be pos or neg and retention time and mass tolerances should be given. The feature table is expected to be pre-formatted.

Usage

annotate_masses(
  features = get_params(step = "annotate_masses")$files$features$prepared,
  output_annotations = get_params(step =
    "annotate_masses")$files$annotations$prepared$structural$ms1,
  output_edges = get_params(step = "annotate_masses")$files$networks$spectral$edges$raw,
  name_source = get_params(step = "annotate_masses")$names$source,
  name_target = get_params(step = "annotate_masses")$names$target,
  library = get_params(step = "annotate_masses")$files$libraries$sop$merged$keys,
  str_stereo = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$npc,
  adducts_list = get_params(step = "annotate_masses")$ms$adducts,
  clusters_list = get_params(step = "annotate_masses")$ms$clusters,
  neutral_losses_list = get_params(step = "annotate_masses")$ms$neutral_losses,
  ms_mode = get_params(step = "annotate_masses")$ms$polarity,
  tolerance_ppm = get_params(step = "annotate_masses")$ms$tolerances$mass$ppm$ms1,
  tolerance_rt = get_params(step = "annotate_masses")$ms$tolerances$rt$adducts
)

Arguments

features

Table containing your previous annotation to complement

output_annotations

Output for mass based structural annotations

output_edges

Output for mass based edges

name_source

Name of the source features column

name_target

Name of the target features column

library

Library containing the keys

str_stereo

File containing structures stereo

str_met

File containing structures metadata

str_nam

File containing structures names

str_tax_cla

File containing Classyfire taxonomy

str_tax_npc

File containing NPClassifier taxonomy

adducts_list

List of adducts to be used

clusters_list

List of clusters to be used

neutral_losses_list

List of neutral losses to be used

ms_mode

Ionization mode. Must be 'pos' or 'neg'

tolerance_ppm

Tolerance to perform annotation. Should be <= 10 ppm

tolerance_rt

Tolerance to group adducts. Should be <= 0.05 minutes

Value

The path to the files containing MS1 annotations and edges

Examples

github <- "https://raw.githubusercontent.com/"
repo <- "taxonomicallyinformedannotation/tima-example-files/"
data_interim <- "main/data/interim/"
dir <- paste0(github, repo, data_interim)
annotate_masses(
  features = paste0(dir, "features/example_features.tsv"),
  library = paste0(dir, "libraries/sop/merged/keys.tsv"),
  str_stereo = paste0(dir, "libraries/sop/merged/structures/stereo.tsv"),
  str_met = paste0(dir, "libraries/sop/merged/structures/metadata.tsv"),
  str_nam = paste0(dir, "libraries/sop/merged/structures/names.tsv"),
  str_tax_cla = paste0(dir, "libraries/sop/merged/structures/taxonomies/classyfire.tsv"),
  str_tax_npc = paste0(dir, "libraries/sop/merged/structures/taxonomies/npc.tsv"),
  adducts_list = list(pos = c("[M+H]+"), neg = c("[M-H]-")),
  clusters_list = list(pos = c("H2O"), neg = c("H2O")),
  neutral_losses_list = c("C6H10O4 (methylpentose/desoxyhexose-H2O)"),
  ms_mode = "pos",
  tolerance_ppm = 10,
  tolerance_rt = 0.01
)
#> 2024-08-12 14:27:51.746131 filtering desired adducts and adding mz tolerance 
#>  
#> 2024-08-12 14:27:51.7561 No previously attributed adducts detected 
#> 2024-08-12 14:27:51.763182 calculating rt tolerance ... 
#>  
#> 2024-08-12 14:27:51.764767 joining within given rt tolerance 
#>  
#> 2024-08-12 14:27:51.773756 adding delta mz tolerance for single charge adducts 
#>  
#> 2024-08-12 14:27:51.778603 forming adducts and clusters 
#>  
#> 2024-08-12 14:27:51.826545 calculating delta mz for single charge adducts and clusters 
#>  
#> 2024-08-12 14:27:51.840629 joining within given delta mz tolerance (neutral losses) 
#>  
#> 2024-08-12 14:27:51.845441 joining within given delta mz tolerance (adducts) 
#>  
#> 2024-08-12 14:27:51.855403 keeping initial and destination feature 
#>  
#> 2024-08-12 14:27:51.862305 joining with initial results (adducts) 
#>  
#> 2024-08-12 14:27:51.865313 joining with initial results (neutral losses) 
#>  
#> 2024-08-12 14:27:51.903774 joining within given mz tol to exact mass library 
#>  
#> 2024-08-12 14:27:51.917141 keeping unique exact masses and molecular formulas 
#>  
#> 2024-08-12 14:27:51.926872 joining exact masses with single charge adducts 
#>  
#> 2024-08-12 14:27:51.926991 Geting back to M 
#>  
#> 2024-08-12 14:27:51.930984 calculating multicharged and in source dimers 
#>  
#> 2024-08-12 14:27:51.940174 Invalid adduct format 
#> 2024-08-12 14:27:51.940297 Returning 0 instead 
#> 2024-08-12 14:27:51.940452 Invalid adduct format 
#> 2024-08-12 14:27:51.940521 Returning 0 instead 
#> 2024-08-12 14:27:51.940657 Invalid adduct format 
#> 2024-08-12 14:27:51.940722 Returning 0 instead 
#> 2024-08-12 14:27:51.940857 Invalid adduct format 
#> 2024-08-12 14:27:51.940955 Returning 0 instead 
#> 2024-08-12 14:27:51.941135 Invalid adduct format 
#> 2024-08-12 14:27:51.941242 Returning 0 instead 
#> 2024-08-12 14:27:51.941395 Invalid adduct format 
#> 2024-08-12 14:27:51.941462 Returning 0 instead 
#> 2024-08-12 14:27:51.94159 Invalid adduct format 
#> 2024-08-12 14:27:51.941654 Returning 0 instead 
#> 2024-08-12 14:27:51.941784 Invalid adduct format 
#> 2024-08-12 14:27:51.941851 Returning 0 instead 
#> 2024-08-12 14:27:51.941977 Invalid adduct format 
#> 2024-08-12 14:27:51.942045 Returning 0 instead 
#> 2024-08-12 14:27:51.942168 Invalid adduct format 
#> 2024-08-12 14:27:51.942232 Returning 0 instead 
#> 2024-08-12 14:27:51.942356 Invalid adduct format 
#> 2024-08-12 14:27:51.942418 Returning 0 instead 
#> 2024-08-12 14:27:51.94254 Invalid adduct format 
#> 2024-08-12 14:27:51.942608 Returning 0 instead 
#> 2024-08-12 14:27:51.942733 Invalid adduct format 
#> 2024-08-12 14:27:51.942798 Returning 0 instead 
#> 2024-08-12 14:27:51.942921 Invalid adduct format 
#> 2024-08-12 14:27:51.942987 Returning 0 instead 
#> 2024-08-12 14:27:51.943112 Invalid adduct format 
#> 2024-08-12 14:27:51.943175 Returning 0 instead 
#> 2024-08-12 14:27:51.943297 Invalid adduct format 
#> 2024-08-12 14:27:51.943359 Returning 0 instead 
#> 2024-08-12 14:27:51.943481 Invalid adduct format 
#> 2024-08-12 14:27:51.943546 Returning 0 instead 
#> 2024-08-12 14:27:51.943666 Invalid adduct format 
#> 2024-08-12 14:27:51.943753 Returning 0 instead 
#> 2024-08-12 14:27:51.943894 Invalid adduct format 
#> 2024-08-12 14:27:51.94396 Returning 0 instead 
#> 2024-08-12 14:27:51.944084 Invalid adduct format 
#> 2024-08-12 14:27:51.944148 Returning 0 instead 
#> 2024-08-12 14:27:51.94427 Invalid adduct format 
#> 2024-08-12 14:27:51.94433 Returning 0 instead 
#> 2024-08-12 14:27:51.944454 Invalid adduct format 
#> 2024-08-12 14:27:51.944515 Returning 0 instead 
#> 2024-08-12 14:27:51.944637 Invalid adduct format 
#> 2024-08-12 14:27:51.94471 Returning 0 instead 
#> 2024-08-12 14:27:51.944835 Invalid adduct format 
#> 2024-08-12 14:27:51.944906 Returning 0 instead 
#> 2024-08-12 14:27:51.945031 Invalid adduct format 
#> 2024-08-12 14:27:51.945096 Returning 0 instead 
#> 2024-08-12 14:27:51.94522 Invalid adduct format 
#> 2024-08-12 14:27:51.945285 Returning 0 instead 
#> 2024-08-12 14:27:51.945406 Invalid adduct format 
#> 2024-08-12 14:27:51.945468 Returning 0 instead 
#> 2024-08-12 14:27:51.94559 Invalid adduct format 
#> 2024-08-12 14:27:51.945653 Returning 0 instead 
#> 2024-08-12 14:27:51.945775 Invalid adduct format 
#> 2024-08-12 14:27:51.945837 Returning 0 instead 
#> 2024-08-12 14:27:51.945958 Invalid adduct format 
#> 2024-08-12 14:27:51.94602 Returning 0 instead 
#> 2024-08-12 14:27:51.946147 Invalid adduct format 
#> 2024-08-12 14:27:51.946214 Returning 0 instead 
#> 2024-08-12 14:27:51.946346 Invalid adduct format 
#> 2024-08-12 14:27:51.946409 Returning 0 instead 
#> 2024-08-12 14:27:51.946533 Invalid adduct format 
#> 2024-08-12 14:27:51.946596 Returning 0 instead 
#> 2024-08-12 14:27:51.946719 Invalid adduct format 
#> 2024-08-12 14:27:51.946782 Returning 0 instead 
#> 2024-08-12 14:27:51.946907 Invalid adduct format 
#> 2024-08-12 14:27:51.946968 Returning 0 instead 
#> 2024-08-12 14:27:51.947092 Invalid adduct format 
#> 2024-08-12 14:27:51.947156 Returning 0 instead 
#> 2024-08-12 14:27:51.947281 Invalid adduct format 
#> 2024-08-12 14:27:51.947344 Returning 0 instead 
#> 2024-08-12 14:27:51.947469 Invalid adduct format 
#> 2024-08-12 14:27:51.947531 Returning 0 instead 
#> 2024-08-12 14:27:51.947651 Invalid adduct format 
#> 2024-08-12 14:27:51.94773 Returning 0 instead 
#> 2024-08-12 14:27:51.947862 Invalid adduct format 
#> 2024-08-12 14:27:51.947927 Returning 0 instead 
#> 2024-08-12 14:27:51.948052 Invalid adduct format 
#> 2024-08-12 14:27:51.948114 Returning 0 instead 
#> 2024-08-12 14:27:51.948235 Invalid adduct format 
#> 2024-08-12 14:27:51.948297 Returning 0 instead 
#> 2024-08-12 14:27:51.94842 Invalid adduct format 
#> 2024-08-12 14:27:51.948483 Returning 0 instead 
#> 2024-08-12 14:27:51.948603 Invalid adduct format 
#> 2024-08-12 14:27:51.948667 Returning 0 instead 
#> 2024-08-12 14:27:51.94879 Invalid adduct format 
#> 2024-08-12 14:27:51.94886 Returning 0 instead 
#> 2024-08-12 14:27:51.948989 Invalid adduct format 
#> 2024-08-12 14:27:51.949055 Returning 0 instead 
#> 2024-08-12 14:27:51.94918 Invalid adduct format 
#> 2024-08-12 14:27:51.949246 Returning 0 instead 
#> 2024-08-12 14:27:51.949372 Invalid adduct format 
#> 2024-08-12 14:27:51.94944 Returning 0 instead 
#> 2024-08-12 14:27:51.949567 Invalid adduct format 
#> 2024-08-12 14:27:51.949636 Returning 0 instead 
#> 2024-08-12 14:27:51.949761 Invalid adduct format 
#> 2024-08-12 14:27:51.949826 Returning 0 instead 
#> 2024-08-12 14:27:51.94996 Invalid adduct format 
#> 2024-08-12 14:27:51.950025 Returning 0 instead 
#> 2024-08-12 14:27:51.950155 Invalid adduct format 
#> 2024-08-12 14:27:51.950224 Returning 0 instead 
#> 2024-08-12 14:27:51.950348 Invalid adduct format 
#> 2024-08-12 14:27:51.950414 Returning 0 instead 
#> 2024-08-12 14:27:51.950539 Invalid adduct format 
#> 2024-08-12 14:27:51.950603 Returning 0 instead 
#> 2024-08-12 14:27:51.950725 Invalid adduct format 
#> 2024-08-12 14:27:51.950788 Returning 0 instead 
#> 2024-08-12 14:27:51.950914 Invalid adduct format 
#> 2024-08-12 14:27:51.95098 Returning 0 instead 
#> 2024-08-12 14:27:51.951103 Invalid adduct format 
#> 2024-08-12 14:27:51.951166 Returning 0 instead 
#> 2024-08-12 14:27:51.951289 Invalid adduct format 
#> 2024-08-12 14:27:51.951351 Returning 0 instead 
#> 2024-08-12 14:27:51.951476 Invalid adduct format 
#> 2024-08-12 14:27:51.951539 Returning 0 instead 
#> 2024-08-12 14:27:51.951663 Invalid adduct format 
#> 2024-08-12 14:27:51.951745 Returning 0 instead 
#> 2024-08-12 14:27:51.951883 Invalid adduct format 
#> 2024-08-12 14:27:51.951948 Returning 0 instead 
#> 2024-08-12 14:27:51.952075 Invalid adduct format 
#> 2024-08-12 14:27:51.952141 Returning 0 instead 
#> 2024-08-12 14:27:51.952264 Invalid adduct format 
#> 2024-08-12 14:27:51.952331 Returning 0 instead 
#> 2024-08-12 14:27:51.952455 Invalid adduct format 
#> 2024-08-12 14:27:51.952521 Returning 0 instead 
#> 2024-08-12 14:27:51.952645 Invalid adduct format 
#> 2024-08-12 14:27:51.95271 Returning 0 instead 
#> 2024-08-12 14:27:51.952832 Invalid adduct format 
#> 2024-08-12 14:27:51.952897 Returning 0 instead 
#> 2024-08-12 14:27:51.95302 Invalid adduct format 
#> 2024-08-12 14:27:51.953086 Returning 0 instead 
#> 2024-08-12 14:27:51.953208 Invalid adduct format 
#> 2024-08-12 14:27:51.953274 Returning 0 instead 
#> 2024-08-12 14:27:51.953396 Invalid adduct format 
#> 2024-08-12 14:27:51.953461 Returning 0 instead 
#> 2024-08-12 14:27:51.953587 Invalid adduct format 
#> 2024-08-12 14:27:51.953651 Returning 0 instead 
#> 2024-08-12 14:27:51.953774 Invalid adduct format 
#> 2024-08-12 14:27:51.953839 Returning 0 instead 
#> 2024-08-12 14:27:51.953964 Invalid adduct format 
#> 2024-08-12 14:27:51.954029 Returning 0 instead 
#> 2024-08-12 14:27:51.954151 Invalid adduct format 
#> 2024-08-12 14:27:51.954214 Returning 0 instead 
#> 2024-08-12 14:27:51.954335 Invalid adduct format 
#> 2024-08-12 14:27:51.954399 Returning 0 instead 
#> 2024-08-12 14:27:51.954521 Invalid adduct format 
#> 2024-08-12 14:27:51.954583 Returning 0 instead 
#> 2024-08-12 14:27:51.954713 Invalid adduct format 
#> 2024-08-12 14:27:51.954782 Returning 0 instead 
#> 2024-08-12 14:27:51.954907 Invalid adduct format 
#> 2024-08-12 14:27:51.954971 Returning 0 instead 
#> 2024-08-12 14:27:51.955096 Invalid adduct format 
#> 2024-08-12 14:27:51.955161 Returning 0 instead 
#> 2024-08-12 14:27:51.955285 Invalid adduct format 
#> 2024-08-12 14:27:51.955347 Returning 0 instead 
#> 2024-08-12 14:27:51.955469 Invalid adduct format 
#> 2024-08-12 14:27:51.955532 Returning 0 instead 
#> 2024-08-12 14:27:51.955653 Invalid adduct format 
#> 2024-08-12 14:27:51.955732 Returning 0 instead 
#> 2024-08-12 14:27:51.955865 Invalid adduct format 
#> 2024-08-12 14:27:51.955928 Returning 0 instead 
#> 2024-08-12 14:27:51.956051 Invalid adduct format 
#> 2024-08-12 14:27:51.956118 Returning 0 instead 
#> 2024-08-12 14:27:51.956249 Invalid adduct format 
#> 2024-08-12 14:27:51.956315 Returning 0 instead 
#> 2024-08-12 14:27:51.956446 Invalid adduct format 
#> 2024-08-12 14:27:51.956512 Returning 0 instead 
#> 2024-08-12 14:27:51.956632 Invalid adduct format 
#> 2024-08-12 14:27:51.956695 Returning 0 instead 
#> 2024-08-12 14:27:51.956817 Invalid adduct format 
#> 2024-08-12 14:27:51.956881 Returning 0 instead 
#> 2024-08-12 14:27:51.957014 Invalid adduct format 
#> 2024-08-12 14:27:51.95708 Returning 0 instead 
#> 2024-08-12 14:27:51.957205 Invalid adduct format 
#> 2024-08-12 14:27:51.957272 Returning 0 instead 
#> 2024-08-12 14:27:51.957396 Invalid adduct format 
#> 2024-08-12 14:27:51.957458 Returning 0 instead 
#> 2024-08-12 14:27:51.957582 Invalid adduct format 
#> 2024-08-12 14:27:51.957647 Returning 0 instead 
#> 2024-08-12 14:27:51.957771 Invalid adduct format 
#> 2024-08-12 14:27:51.957835 Returning 0 instead 
#> 2024-08-12 14:27:51.957954 Invalid adduct format 
#> 2024-08-12 14:27:51.958018 Returning 0 instead 
#> 2024-08-12 14:27:51.958137 Invalid adduct format 
#> 2024-08-12 14:27:51.958201 Returning 0 instead 
#> 2024-08-12 14:27:51.958321 Invalid adduct format 
#> 2024-08-12 14:27:51.958385 Returning 0 instead 
#> 2024-08-12 14:27:51.958507 Invalid adduct format 
#> 2024-08-12 14:27:51.95857 Returning 0 instead 
#> 2024-08-12 14:27:51.958694 Invalid adduct format 
#> 2024-08-12 14:27:51.958757 Returning 0 instead 
#> 2024-08-12 14:27:51.958879 Invalid adduct format 
#> 2024-08-12 14:27:51.958941 Returning 0 instead 
#> 2024-08-12 14:27:51.959063 Invalid adduct format 
#> 2024-08-12 14:27:51.959127 Returning 0 instead 
#> 2024-08-12 14:27:51.95925 Invalid adduct format 
#> 2024-08-12 14:27:51.959315 Returning 0 instead 
#> 2024-08-12 14:27:51.959438 Invalid adduct format 
#> 2024-08-12 14:27:51.959502 Returning 0 instead 
#> 2024-08-12 14:27:51.963658 joining within given rt tolerance 
#>  
#> 2024-08-12 14:27:51.966051 joining within given mz tol and filtering possible adducts 
#>  
#> 2024-08-12 14:27:51.978164 joining single adducts, in source dimers, and multicharged 
#>  
#> 2024-08-12 14:27:51.983287 adding chemical classification 
#> 2024-08-12 14:27:51.995351 MS1 annotation led to 
#>  0 annotations , on 
#>  0 features 
#> Warning: cannot open file '/home/runner/work/_temp/Library/timaparams/default/annotate_masses.yaml': No such file or directory
#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection