Change Params Small

This function helps changing convenience parameters.

Usage

change_params_small(
  fil_pat = NULL,
  fil_fea_raw = NULL,
  fil_met_raw = NULL,
  fil_sir_raw = NULL,
  fil_spe_raw = NULL,
  ms_pol = NULL,
  org_tax = NULL,
  hig_con = NULL,
  summarize = NULL
)

Arguments

fil_pat

The pattern identifying your whole job. You can put whatever you want. STRING

fil_fea_raw

The path to the file containing your features' intensities. Can be generated by mzmine, SLAW, or SIRIUS. STRING

fil_met_raw

The path to the file containing your metadata. If your experiment contains a single taxon, you can provide it below instead. STRING

fil_sir_raw

The directory containing the sirius annotations. STRING

fil_spe_raw

The path to the file containing your features' spectra. Can contain MS1 and/or MS2 spectra. STRING

ms_pol

The polarity used. Must be either "pos" or "neg". STRING

org_tax

If your experiment contains a single taxon, its scientific name. "Homo sapiens". STRING

hig_con

Filter high confidence candidates only. BOOLEAN

summarize

Summarize all candidates per feature to a single row. BOOLEAN

Value

YAML file with changed parameters.

Examples

if (FALSE) { # \dontrun{
copy_backbone()
tima::change_params_small(
  fil_pat = "myExamplePattern",
  fil_fea_raw = "myExampleDir/myExampleFeatures.csv",
  fil_met_raw = "myExampleDir2SomeWhereElse/myOptionalMetadata.tsv",
  fil_sir_raw = "myExampleDir3/myAwesomeSiriusProject.zip",
  fil_spe_raw = "myBeautifulSpectra.mgf",
  ms_pol = "pos",
  org_tax = "Gentiana lutea",
  hig_con = TRUE,
  summarize = FALSE
)
} # }