Clean chemo

This function cleans the results obtained after chemical weighting

Usage

clean_chemo(
  annot_table_wei_chemo = get("annot_table_wei_chemo", envir = parent.frame()),
  components_table = get("components_table", envir = parent.frame()),
  features_table = get("features_table", envir = parent.frame()),
  structure_organism_pairs_table = get("structure_organism_pairs_table", envir =
    parent.frame()),
  candidates_final = get("candidates_final", envir = parent.frame()),
  minimal_ms1_bio = get("minimal_ms1_bio", envir = parent.frame()),
  minimal_ms1_chemo = get("minimal_ms1_chemo", envir = parent.frame()),
  minimal_ms1_condition = get("minimal_ms1_condition", envir = parent.frame()),
  compounds_names = get("compounds_names", envir = parent.frame()),
  high_confidence = get("high_confidence", envir = parent.frame()),
  remove_ties = get("remove_ties", envir = parent.frame()),
  summarize = get("summarize", envir = parent.frame())
)

Arguments

annot_table_wei_chemo

Table containing your chemically weighted annotation

components_table

Prepared components file

features_table

Prepared features file

structure_organism_pairs_table

Table containing the structure - organism pairs

candidates_final

Number of final candidates to keep

minimal_ms1_bio

Minimal biological score to keep MS1 based annotation

minimal_ms1_chemo

Minimal chemical score to keep MS1 based annotation

minimal_ms1_condition

Condition to be used. Must be "OR" or "AND".

compounds_names

Report compounds names. Can be very large. BOOLEAN

high_confidence

Report high confidence candidates only. BOOLEAN

remove_ties

Remove ties. BOOLEAN

summarize

Boolean. summarize results (1 row per feature)

Value

A table containing the chemically weighted annotation where only a given number of initial candidates are kept

See also

Examples

NULL
#> NULL