Annotate masses

This function annotates a feature table based on exact mass match. It requires a structural library, its metadata, and lists of adducts, clusters, and neutral losses to be considered. The polarity has to be pos or neg and retention time and mass tolerances should be given. The feature table is expected to be pre-formatted.

Usage

annotate_masses(
  features = get_params(step = "annotate_masses")$files$features$prepared,
  output_annotations = get_params(step =
    "annotate_masses")$files$annotations$prepared$structural$ms1,
  output_edges = get_params(step = "annotate_masses")$files$networks$spectral$edges$raw,
  name_source = get_params(step = "annotate_masses")$names$source,
  name_target = get_params(step = "annotate_masses")$names$target,
  library = get_params(step = "annotate_masses")$files$libraries$sop$merged$keys,
  str_stereo = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$npc,
  adducts_list = get_params(step = "annotate_masses")$ms$adducts,
  clusters_list = get_params(step = "annotate_masses")$ms$clusters,
  neutral_losses_list = get_params(step = "annotate_masses")$ms$neutral_losses,
  ms_mode = get_params(step = "annotate_masses")$ms$polarity,
  tolerance_ppm = get_params(step = "annotate_masses")$ms$tolerances$mass$ppm$ms1,
  tolerance_rt = get_params(step = "annotate_masses")$ms$tolerances$rt$adducts
)

Arguments

features

Table containing your previous annotation to complement

output_annotations

Output for mass based structural annotations

output_edges

Output for mass based edges

name_source

Name of the source features column

name_target

Name of the target features column

library

Library containing the keys

str_stereo

File containing structures stereo

str_met

File containing structures metadata

str_nam

File containing structures names

str_tax_cla

File containing Classyfire taxonomy

str_tax_npc

File containing NPClassifier taxonomy

adducts_list

List of adducts to be used

clusters_list

List of clusters to be used

neutral_losses_list

List of neutral losses to be used

ms_mode

Ionization mode. Must be 'pos' or 'neg'

tolerance_ppm

Tolerance to perform annotation. Should be <= 10 ppm

tolerance_rt

Tolerance to group adducts. Should be <= 0.05 minutes

Value

The path to the files containing MS1 annotations and edges

Examples

github <- "https://raw.githubusercontent.com/"
repo <- "taxonomicallyinformedannotation/tima-example-files/"
data_interim <- "main/data/interim/"
dir <- paste0(github, repo, data_interim)
annotate_masses(
  features = paste0(dir, "features/example_features.tsv"),
  library = paste0(dir, "libraries/sop/merged/keys.tsv"),
  str_stereo = paste0(dir, "libraries/sop/merged/structures/stereo.tsv"),
  str_met = paste0(dir, "libraries/sop/merged/structures/metadata.tsv"),
  str_nam = paste0(dir, "libraries/sop/merged/structures/names.tsv"),
  str_tax_cla = paste0(dir, "libraries/sop/merged/structures/taxonomies/classyfire.tsv"),
  str_tax_npc = paste0(dir, "libraries/sop/merged/structures/taxonomies/npc.tsv"),
  adducts_list = list(pos = c("[M+H]+"), neg = c("[M-H]-")),
  clusters_list = list(pos = c("H2O"), neg = c("H2O")),
  neutral_losses_list = c("C6H10O4 (methylpentose/desoxyhexose-H2O)"),
  ms_mode = "pos",
  tolerance_ppm = 10,
  tolerance_rt = 0.01
)
#> 2024-08-12 14:44:46.031659 filtering desired adducts and adding mz tolerance 
#>  
#> 2024-08-12 14:44:46.042691 No previously attributed adducts detected 
#> 2024-08-12 14:44:46.050586 calculating rt tolerance ... 
#>  
#> 2024-08-12 14:44:46.05226 joining within given rt tolerance 
#>  
#> 2024-08-12 14:44:46.062317 adding delta mz tolerance for single charge adducts 
#>  
#> 2024-08-12 14:44:46.067605 forming adducts and clusters 
#>  
#> 2024-08-12 14:44:46.089156 calculating delta mz for single charge adducts and clusters 
#>  
#> 2024-08-12 14:44:46.10482 joining within given delta mz tolerance (neutral losses) 
#>  
#> 2024-08-12 14:44:46.110221 joining within given delta mz tolerance (adducts) 
#>  
#> 2024-08-12 14:44:46.121525 keeping initial and destination feature 
#>  
#> 2024-08-12 14:44:46.128989 joining with initial results (adducts) 
#>  
#> 2024-08-12 14:44:46.132275 joining with initial results (neutral losses) 
#>  
#> 2024-08-12 14:44:46.173014 joining within given mz tol to exact mass library 
#>  
#> 2024-08-12 14:44:46.187848 keeping unique exact masses and molecular formulas 
#>  
#> 2024-08-12 14:44:46.198132 joining exact masses with single charge adducts 
#>  
#> 2024-08-12 14:44:46.198258 Geting back to M 
#>  
#> 2024-08-12 14:44:46.202449 calculating multicharged and in source dimers 
#>  
#> 2024-08-12 14:44:46.212482 Invalid adduct format 
#> 2024-08-12 14:44:46.212618 Returning 0 instead 
#> 2024-08-12 14:44:46.212775 Invalid adduct format 
#> 2024-08-12 14:44:46.212848 Returning 0 instead 
#> 2024-08-12 14:44:46.212988 Invalid adduct format 
#> 2024-08-12 14:44:46.213057 Returning 0 instead 
#> 2024-08-12 14:44:46.213191 Invalid adduct format 
#> 2024-08-12 14:44:46.213261 Returning 0 instead 
#> 2024-08-12 14:44:46.213391 Invalid adduct format 
#> 2024-08-12 14:44:46.213458 Returning 0 instead 
#> 2024-08-12 14:44:46.213629 Invalid adduct format 
#> 2024-08-12 14:44:46.213727 Returning 0 instead 
#> 2024-08-12 14:44:46.213866 Invalid adduct format 
#> 2024-08-12 14:44:46.213934 Returning 0 instead 
#> 2024-08-12 14:44:46.214064 Invalid adduct format 
#> 2024-08-12 14:44:46.214127 Returning 0 instead 
#> 2024-08-12 14:44:46.214252 Invalid adduct format 
#> 2024-08-12 14:44:46.214317 Returning 0 instead 
#> 2024-08-12 14:44:46.21444 Invalid adduct format 
#> 2024-08-12 14:44:46.214502 Returning 0 instead 
#> 2024-08-12 14:44:46.214624 Invalid adduct format 
#> 2024-08-12 14:44:46.214694 Returning 0 instead 
#> 2024-08-12 14:44:46.214821 Invalid adduct format 
#> 2024-08-12 14:44:46.21489 Returning 0 instead 
#> 2024-08-12 14:44:46.21502 Invalid adduct format 
#> 2024-08-12 14:44:46.215087 Returning 0 instead 
#> 2024-08-12 14:44:46.215213 Invalid adduct format 
#> 2024-08-12 14:44:46.215277 Returning 0 instead 
#> 2024-08-12 14:44:46.215401 Invalid adduct format 
#> 2024-08-12 14:44:46.215467 Returning 0 instead 
#> 2024-08-12 14:44:46.21559 Invalid adduct format 
#> 2024-08-12 14:44:46.215653 Returning 0 instead 
#> 2024-08-12 14:44:46.215772 Invalid adduct format 
#> 2024-08-12 14:44:46.215833 Returning 0 instead 
#> 2024-08-12 14:44:46.215959 Invalid adduct format 
#> 2024-08-12 14:44:46.216026 Returning 0 instead 
#> 2024-08-12 14:44:46.21615 Invalid adduct format 
#> 2024-08-12 14:44:46.216213 Returning 0 instead 
#> 2024-08-12 14:44:46.21634 Invalid adduct format 
#> 2024-08-12 14:44:46.216403 Returning 0 instead 
#> 2024-08-12 14:44:46.216524 Invalid adduct format 
#> 2024-08-12 14:44:46.21659 Returning 0 instead 
#> 2024-08-12 14:44:46.216744 Invalid adduct format 
#> 2024-08-12 14:44:46.216813 Returning 0 instead 
#> 2024-08-12 14:44:46.216941 Invalid adduct format 
#> 2024-08-12 14:44:46.217005 Returning 0 instead 
#> 2024-08-12 14:44:46.217131 Invalid adduct format 
#> 2024-08-12 14:44:46.217203 Returning 0 instead 
#> 2024-08-12 14:44:46.21733 Invalid adduct format 
#> 2024-08-12 14:44:46.217394 Returning 0 instead 
#> 2024-08-12 14:44:46.217521 Invalid adduct format 
#> 2024-08-12 14:44:46.217608 Returning 0 instead 
#> 2024-08-12 14:44:46.217736 Invalid adduct format 
#> 2024-08-12 14:44:46.217802 Returning 0 instead 
#> 2024-08-12 14:44:46.217925 Invalid adduct format 
#> 2024-08-12 14:44:46.217989 Returning 0 instead 
#> 2024-08-12 14:44:46.21811 Invalid adduct format 
#> 2024-08-12 14:44:46.218171 Returning 0 instead 
#> 2024-08-12 14:44:46.218297 Invalid adduct format 
#> 2024-08-12 14:44:46.218364 Returning 0 instead 
#> 2024-08-12 14:44:46.218499 Invalid adduct format 
#> 2024-08-12 14:44:46.21857 Returning 0 instead 
#> 2024-08-12 14:44:46.218698 Invalid adduct format 
#> 2024-08-12 14:44:46.218768 Returning 0 instead 
#> 2024-08-12 14:44:46.218895 Invalid adduct format 
#> 2024-08-12 14:44:46.218963 Returning 0 instead 
#> 2024-08-12 14:44:46.219088 Invalid adduct format 
#> 2024-08-12 14:44:46.219161 Returning 0 instead 
#> 2024-08-12 14:44:46.219285 Invalid adduct format 
#> 2024-08-12 14:44:46.219349 Returning 0 instead 
#> 2024-08-12 14:44:46.219474 Invalid adduct format 
#> 2024-08-12 14:44:46.219538 Returning 0 instead 
#> 2024-08-12 14:44:46.219665 Invalid adduct format 
#> 2024-08-12 14:44:46.219734 Returning 0 instead 
#> 2024-08-12 14:44:46.219864 Invalid adduct format 
#> 2024-08-12 14:44:46.219931 Returning 0 instead 
#> 2024-08-12 14:44:46.220056 Invalid adduct format 
#> 2024-08-12 14:44:46.220129 Returning 0 instead 
#> 2024-08-12 14:44:46.220257 Invalid adduct format 
#> 2024-08-12 14:44:46.220323 Returning 0 instead 
#> 2024-08-12 14:44:46.220449 Invalid adduct format 
#> 2024-08-12 14:44:46.220514 Returning 0 instead 
#> 2024-08-12 14:44:46.220638 Invalid adduct format 
#> 2024-08-12 14:44:46.220702 Returning 0 instead 
#> 2024-08-12 14:44:46.220826 Invalid adduct format 
#> 2024-08-12 14:44:46.220893 Returning 0 instead 
#> 2024-08-12 14:44:46.221024 Invalid adduct format 
#> 2024-08-12 14:44:46.221103 Returning 0 instead 
#> 2024-08-12 14:44:46.221236 Invalid adduct format 
#> 2024-08-12 14:44:46.221303 Returning 0 instead 
#> 2024-08-12 14:44:46.221433 Invalid adduct format 
#> 2024-08-12 14:44:46.221501 Returning 0 instead 
#> 2024-08-12 14:44:46.221667 Invalid adduct format 
#> 2024-08-12 14:44:46.221745 Returning 0 instead 
#> 2024-08-12 14:44:46.221872 Invalid adduct format 
#> 2024-08-12 14:44:46.221945 Returning 0 instead 
#> 2024-08-12 14:44:46.222078 Invalid adduct format 
#> 2024-08-12 14:44:46.22215 Returning 0 instead 
#> 2024-08-12 14:44:46.222279 Invalid adduct format 
#> 2024-08-12 14:44:46.222341 Returning 0 instead 
#> 2024-08-12 14:44:46.222463 Invalid adduct format 
#> 2024-08-12 14:44:46.222533 Returning 0 instead 
#> 2024-08-12 14:44:46.222659 Invalid adduct format 
#> 2024-08-12 14:44:46.222724 Returning 0 instead 
#> 2024-08-12 14:44:46.222853 Invalid adduct format 
#> 2024-08-12 14:44:46.22292 Returning 0 instead 
#> 2024-08-12 14:44:46.223043 Invalid adduct format 
#> 2024-08-12 14:44:46.22311 Returning 0 instead 
#> 2024-08-12 14:44:46.223236 Invalid adduct format 
#> 2024-08-12 14:44:46.223301 Returning 0 instead 
#> 2024-08-12 14:44:46.223425 Invalid adduct format 
#> 2024-08-12 14:44:46.223489 Returning 0 instead 
#> 2024-08-12 14:44:46.223622 Invalid adduct format 
#> 2024-08-12 14:44:46.223689 Returning 0 instead 
#> 2024-08-12 14:44:46.223813 Invalid adduct format 
#> 2024-08-12 14:44:46.223878 Returning 0 instead 
#> 2024-08-12 14:44:46.224002 Invalid adduct format 
#> 2024-08-12 14:44:46.224069 Returning 0 instead 
#> 2024-08-12 14:44:46.224191 Invalid adduct format 
#> 2024-08-12 14:44:46.224254 Returning 0 instead 
#> 2024-08-12 14:44:46.224379 Invalid adduct format 
#> 2024-08-12 14:44:46.224444 Returning 0 instead 
#> 2024-08-12 14:44:46.224566 Invalid adduct format 
#> 2024-08-12 14:44:46.224634 Returning 0 instead 
#> 2024-08-12 14:44:46.224759 Invalid adduct format 
#> 2024-08-12 14:44:46.224825 Returning 0 instead 
#> 2024-08-12 14:44:46.224951 Invalid adduct format 
#> 2024-08-12 14:44:46.225021 Returning 0 instead 
#> 2024-08-12 14:44:46.225147 Invalid adduct format 
#> 2024-08-12 14:44:46.225215 Returning 0 instead 
#> 2024-08-12 14:44:46.22534 Invalid adduct format 
#> 2024-08-12 14:44:46.225405 Returning 0 instead 
#> 2024-08-12 14:44:46.225529 Invalid adduct format 
#> 2024-08-12 14:44:46.225622 Returning 0 instead 
#> 2024-08-12 14:44:46.225753 Invalid adduct format 
#> 2024-08-12 14:44:46.225824 Returning 0 instead 
#> 2024-08-12 14:44:46.225948 Invalid adduct format 
#> 2024-08-12 14:44:46.226013 Returning 0 instead 
#> 2024-08-12 14:44:46.226137 Invalid adduct format 
#> 2024-08-12 14:44:46.226203 Returning 0 instead 
#> 2024-08-12 14:44:46.226332 Invalid adduct format 
#> 2024-08-12 14:44:46.226397 Returning 0 instead 
#> 2024-08-12 14:44:46.226525 Invalid adduct format 
#> 2024-08-12 14:44:46.226591 Returning 0 instead 
#> 2024-08-12 14:44:46.226721 Invalid adduct format 
#> 2024-08-12 14:44:46.226813 Returning 0 instead 
#> 2024-08-12 14:44:46.226948 Invalid adduct format 
#> 2024-08-12 14:44:46.227015 Returning 0 instead 
#> 2024-08-12 14:44:46.227142 Invalid adduct format 
#> 2024-08-12 14:44:46.227209 Returning 0 instead 
#> 2024-08-12 14:44:46.227334 Invalid adduct format 
#> 2024-08-12 14:44:46.227399 Returning 0 instead 
#> 2024-08-12 14:44:46.227523 Invalid adduct format 
#> 2024-08-12 14:44:46.227589 Returning 0 instead 
#> 2024-08-12 14:44:46.227716 Invalid adduct format 
#> 2024-08-12 14:44:46.227784 Returning 0 instead 
#> 2024-08-12 14:44:46.22791 Invalid adduct format 
#> 2024-08-12 14:44:46.227974 Returning 0 instead 
#> 2024-08-12 14:44:46.228101 Invalid adduct format 
#> 2024-08-12 14:44:46.228165 Returning 0 instead 
#> 2024-08-12 14:44:46.228293 Invalid adduct format 
#> 2024-08-12 14:44:46.228357 Returning 0 instead 
#> 2024-08-12 14:44:46.228477 Invalid adduct format 
#> 2024-08-12 14:44:46.228544 Returning 0 instead 
#> 2024-08-12 14:44:46.228668 Invalid adduct format 
#> 2024-08-12 14:44:46.228731 Returning 0 instead 
#> 2024-08-12 14:44:46.228853 Invalid adduct format 
#> 2024-08-12 14:44:46.228919 Returning 0 instead 
#> 2024-08-12 14:44:46.22904 Invalid adduct format 
#> 2024-08-12 14:44:46.229103 Returning 0 instead 
#> 2024-08-12 14:44:46.229225 Invalid adduct format 
#> 2024-08-12 14:44:46.229292 Returning 0 instead 
#> 2024-08-12 14:44:46.229417 Invalid adduct format 
#> 2024-08-12 14:44:46.229481 Returning 0 instead 
#> 2024-08-12 14:44:46.229638 Invalid adduct format 
#> 2024-08-12 14:44:46.229712 Returning 0 instead 
#> 2024-08-12 14:44:46.229838 Invalid adduct format 
#> 2024-08-12 14:44:46.229903 Returning 0 instead 
#> 2024-08-12 14:44:46.230024 Invalid adduct format 
#> 2024-08-12 14:44:46.230092 Returning 0 instead 
#> 2024-08-12 14:44:46.230216 Invalid adduct format 
#> 2024-08-12 14:44:46.230282 Returning 0 instead 
#> 2024-08-12 14:44:46.230404 Invalid adduct format 
#> 2024-08-12 14:44:46.230467 Returning 0 instead 
#> 2024-08-12 14:44:46.230589 Invalid adduct format 
#> 2024-08-12 14:44:46.230652 Returning 0 instead 
#> 2024-08-12 14:44:46.230776 Invalid adduct format 
#> 2024-08-12 14:44:46.230839 Returning 0 instead 
#> 2024-08-12 14:44:46.230961 Invalid adduct format 
#> 2024-08-12 14:44:46.231027 Returning 0 instead 
#> 2024-08-12 14:44:46.23116 Invalid adduct format 
#> 2024-08-12 14:44:46.231228 Returning 0 instead 
#> 2024-08-12 14:44:46.23135 Invalid adduct format 
#> 2024-08-12 14:44:46.231413 Returning 0 instead 
#> 2024-08-12 14:44:46.231534 Invalid adduct format 
#> 2024-08-12 14:44:46.2316 Returning 0 instead 
#> 2024-08-12 14:44:46.231724 Invalid adduct format 
#> 2024-08-12 14:44:46.23179 Returning 0 instead 
#> 2024-08-12 14:44:46.231913 Invalid adduct format 
#> 2024-08-12 14:44:46.231982 Returning 0 instead 
#> 2024-08-12 14:44:46.232108 Invalid adduct format 
#> 2024-08-12 14:44:46.232194 Returning 0 instead 
#> 2024-08-12 14:44:46.23666 joining within given rt tolerance 
#>  
#> 2024-08-12 14:44:46.239183 joining within given mz tol and filtering possible adducts 
#>  
#> 2024-08-12 14:44:46.252198 joining single adducts, in source dimers, and multicharged 
#>  
#> 2024-08-12 14:44:46.257406 adding chemical classification 
#> 2024-08-12 14:44:46.27091 MS1 annotation led to 
#>  0 annotations , on 
#>  0 features 
#> Warning: cannot open file '/home/runner/work/_temp/Library/timaparams/default/annotate_masses.yaml': No such file or directory
#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection