Annotate masses

This function annotates a feature table based on exact mass match. It requires a structural library, its metadata, and lists of adducts, clusters, and neutral losses to be considered. The polarity has to be pos or neg and retention time and mass tolerances should be given. The feature table is expected to be pre-formatted.

Usage

annotate_masses(
  features = get_params(step = "annotate_masses")$files$features$prepared,
  output_annotations = get_params(step =
    "annotate_masses")$files$annotations$prepared$structural$ms1,
  output_edges = get_params(step = "annotate_masses")$files$networks$spectral$edges$raw,
  name_source = get_params(step = "annotate_masses")$names$source,
  name_target = get_params(step = "annotate_masses")$names$target,
  library = get_params(step = "annotate_masses")$files$libraries$sop$merged$keys,
  str_stereo = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$npc,
  adducts_list = get_params(step = "annotate_masses")$ms$adducts,
  clusters_list = get_params(step = "annotate_masses")$ms$clusters,
  neutral_losses_list = get_params(step = "annotate_masses")$ms$neutral_losses,
  ms_mode = get_params(step = "annotate_masses")$ms$polarity,
  tolerance_ppm = get_params(step = "annotate_masses")$ms$tolerances$mass$ppm$ms1,
  tolerance_rt = get_params(step = "annotate_masses")$ms$tolerances$rt$adducts
)

Arguments

features

Table containing your previous annotation to complement

output_annotations

Output for mass based structural annotations

output_edges

Output for mass based edges

name_source

Name of the source features column

name_target

Name of the target features column

library

Library containing the keys

str_stereo

File containing structures stereo

str_met

File containing structures metadata

str_nam

File containing structures names

str_tax_cla

File containing Classyfire taxonomy

str_tax_npc

File containing NPClassifier taxonomy

adducts_list

List of adducts to be used

clusters_list

List of clusters to be used

neutral_losses_list

List of neutral losses to be used

ms_mode

Ionization mode. Must be 'pos' or 'neg'

tolerance_ppm

Tolerance to perform annotation. Should be <= 10 ppm

tolerance_rt

Tolerance to group adducts. Should be <= 0.05 minutes

Value

The path to the files containing MS1 annotations and edges

Examples

github <- "https://raw.githubusercontent.com/"
repo <- "taxonomicallyinformedannotation/tima-example-files/"
data_interim <- "main/data/interim/"
dir <- paste0(github, repo, data_interim)
annotate_masses(
  features = paste0(dir, "features/example_features.tsv"),
  library = paste0(dir, "libraries/sop/merged/keys.tsv"),
  str_stereo = paste0(dir, "libraries/sop/merged/structures/stereo.tsv"),
  str_met = paste0(dir, "libraries/sop/merged/structures/metadata.tsv"),
  str_nam = paste0(dir, "libraries/sop/merged/structures/names.tsv"),
  str_tax_cla = paste0(dir, "libraries/sop/merged/structures/taxonomies/classyfire.tsv"),
  str_tax_npc = paste0(dir, "libraries/sop/merged/structures/taxonomies/npc.tsv"),
  adducts_list = list(pos = c("[M+H]+"), neg = c("[M-H]-")),
  clusters_list = list(pos = c("H2O"), neg = c("H2O")),
  neutral_losses_list = c("C6H10O4 (methylpentose/desoxyhexose-H2O)"),
  ms_mode = "pos",
  tolerance_ppm = 10,
  tolerance_rt = 0.01,
  output_annotations = "output_annotations.tsv",
  output_edges = "output_edges.tsv"
)
#> 2024-08-12 17:32:49.222139 filtering desired adducts and adding mz tolerance 
#>  
#> 2024-08-12 17:32:49.231967 No previously attributed adducts detected 
#> 2024-08-12 17:32:49.238775 calculating rt tolerance ... 
#>  
#> 2024-08-12 17:32:49.240348 joining within given rt tolerance 
#>  
#> 2024-08-12 17:32:49.249166 adding delta mz tolerance for single charge adducts 
#>  
#> 2024-08-12 17:32:49.253968 forming adducts and clusters 
#>  
#> 2024-08-12 17:32:49.268111 calculating delta mz for single charge adducts and clusters 
#>  
#> 2024-08-12 17:32:49.281755 joining within given delta mz tolerance (neutral losses) 
#>  
#> 2024-08-12 17:32:49.286662 joining within given delta mz tolerance (adducts) 
#>  
#> 2024-08-12 17:32:49.300397 keeping initial and destination feature 
#>  
#> 2024-08-12 17:32:49.3074 joining with initial results (adducts) 
#>  
#> 2024-08-12 17:32:49.310426 joining with initial results (neutral losses) 
#>  
#> 2024-08-12 17:32:49.361511 joining within given mz tol to exact mass library 
#>  
#> 2024-08-12 17:32:49.36787 keeping unique exact masses and molecular formulas 
#>  
#> 2024-08-12 17:32:49.377483 joining exact masses with single charge adducts 
#>  
#> 2024-08-12 17:32:49.377587 Geting back to M 
#>  
#> 2024-08-12 17:32:49.381236 calculating multicharged and in source dimers 
#>  
#> 2024-08-12 17:32:49.390208 Invalid adduct format 
#> 2024-08-12 17:32:49.390327 Returning 0 instead 
#> 2024-08-12 17:32:49.390481 Invalid adduct format 
#> 2024-08-12 17:32:49.390549 Returning 0 instead 
#> 2024-08-12 17:32:49.390686 Invalid adduct format 
#> 2024-08-12 17:32:49.390761 Returning 0 instead 
#> 2024-08-12 17:32:49.39089 Invalid adduct format 
#> 2024-08-12 17:32:49.390952 Returning 0 instead 
#> 2024-08-12 17:32:49.391082 Invalid adduct format 
#> 2024-08-12 17:32:49.391144 Returning 0 instead 
#> 2024-08-12 17:32:49.39127 Invalid adduct format 
#> 2024-08-12 17:32:49.391332 Returning 0 instead 
#> 2024-08-12 17:32:49.391458 Invalid adduct format 
#> 2024-08-12 17:32:49.391521 Returning 0 instead 
#> 2024-08-12 17:32:49.391672 Invalid adduct format 
#> 2024-08-12 17:32:49.391739 Returning 0 instead 
#> 2024-08-12 17:32:49.391866 Invalid adduct format 
#> 2024-08-12 17:32:49.391929 Returning 0 instead 
#> 2024-08-12 17:32:49.392055 Invalid adduct format 
#> 2024-08-12 17:32:49.392117 Returning 0 instead 
#> 2024-08-12 17:32:49.39224 Invalid adduct format 
#> 2024-08-12 17:32:49.392312 Returning 0 instead 
#> 2024-08-12 17:32:49.392438 Invalid adduct format 
#> 2024-08-12 17:32:49.392505 Returning 0 instead 
#> 2024-08-12 17:32:49.392628 Invalid adduct format 
#> 2024-08-12 17:32:49.392691 Returning 0 instead 
#> 2024-08-12 17:32:49.392813 Invalid adduct format 
#> 2024-08-12 17:32:49.392877 Returning 0 instead 
#> 2024-08-12 17:32:49.393002 Invalid adduct format 
#> 2024-08-12 17:32:49.393065 Returning 0 instead 
#> 2024-08-12 17:32:49.393188 Invalid adduct format 
#> 2024-08-12 17:32:49.39325 Returning 0 instead 
#> 2024-08-12 17:32:49.393371 Invalid adduct format 
#> 2024-08-12 17:32:49.393433 Returning 0 instead 
#> 2024-08-12 17:32:49.393555 Invalid adduct format 
#> 2024-08-12 17:32:49.393618 Returning 0 instead 
#> 2024-08-12 17:32:49.393741 Invalid adduct format 
#> 2024-08-12 17:32:49.393802 Returning 0 instead 
#> 2024-08-12 17:32:49.393924 Invalid adduct format 
#> 2024-08-12 17:32:49.393985 Returning 0 instead 
#> 2024-08-12 17:32:49.394104 Invalid adduct format 
#> 2024-08-12 17:32:49.394164 Returning 0 instead 
#> 2024-08-12 17:32:49.394285 Invalid adduct format 
#> 2024-08-12 17:32:49.394353 Returning 0 instead 
#> 2024-08-12 17:32:49.394475 Invalid adduct format 
#> 2024-08-12 17:32:49.394538 Returning 0 instead 
#> 2024-08-12 17:32:49.394658 Invalid adduct format 
#> 2024-08-12 17:32:49.39472 Returning 0 instead 
#> 2024-08-12 17:32:49.394849 Invalid adduct format 
#> 2024-08-12 17:32:49.394912 Returning 0 instead 
#> 2024-08-12 17:32:49.395033 Invalid adduct format 
#> 2024-08-12 17:32:49.395095 Returning 0 instead 
#> 2024-08-12 17:32:49.395216 Invalid adduct format 
#> 2024-08-12 17:32:49.395281 Returning 0 instead 
#> 2024-08-12 17:32:49.395409 Invalid adduct format 
#> 2024-08-12 17:32:49.395473 Returning 0 instead 
#> 2024-08-12 17:32:49.395618 Invalid adduct format 
#> 2024-08-12 17:32:49.395689 Returning 0 instead 
#> 2024-08-12 17:32:49.395816 Invalid adduct format 
#> 2024-08-12 17:32:49.395881 Returning 0 instead 
#> 2024-08-12 17:32:49.396005 Invalid adduct format 
#> 2024-08-12 17:32:49.396069 Returning 0 instead 
#> 2024-08-12 17:32:49.396196 Invalid adduct format 
#> 2024-08-12 17:32:49.396261 Returning 0 instead 
#> 2024-08-12 17:32:49.396387 Invalid adduct format 
#> 2024-08-12 17:32:49.396454 Returning 0 instead 
#> 2024-08-12 17:32:49.396585 Invalid adduct format 
#> 2024-08-12 17:32:49.396654 Returning 0 instead 
#> 2024-08-12 17:32:49.39678 Invalid adduct format 
#> 2024-08-12 17:32:49.396846 Returning 0 instead 
#> 2024-08-12 17:32:49.396967 Invalid adduct format 
#> 2024-08-12 17:32:49.397031 Returning 0 instead 
#> 2024-08-12 17:32:49.397156 Invalid adduct format 
#> 2024-08-12 17:32:49.397224 Returning 0 instead 
#> 2024-08-12 17:32:49.397346 Invalid adduct format 
#> 2024-08-12 17:32:49.397409 Returning 0 instead 
#> 2024-08-12 17:32:49.397528 Invalid adduct format 
#> 2024-08-12 17:32:49.397593 Returning 0 instead 
#> 2024-08-12 17:32:49.397714 Invalid adduct format 
#> 2024-08-12 17:32:49.397776 Returning 0 instead 
#> 2024-08-12 17:32:49.397895 Invalid adduct format 
#> 2024-08-12 17:32:49.397956 Returning 0 instead 
#> 2024-08-12 17:32:49.398077 Invalid adduct format 
#> 2024-08-12 17:32:49.398142 Returning 0 instead 
#> 2024-08-12 17:32:49.398271 Invalid adduct format 
#> 2024-08-12 17:32:49.398335 Returning 0 instead 
#> 2024-08-12 17:32:49.398457 Invalid adduct format 
#> 2024-08-12 17:32:49.398517 Returning 0 instead 
#> 2024-08-12 17:32:49.398638 Invalid adduct format 
#> 2024-08-12 17:32:49.398703 Returning 0 instead 
#> 2024-08-12 17:32:49.398827 Invalid adduct format 
#> 2024-08-12 17:32:49.39889 Returning 0 instead 
#> 2024-08-12 17:32:49.399015 Invalid adduct format 
#> 2024-08-12 17:32:49.39908 Returning 0 instead 
#> 2024-08-12 17:32:49.399203 Invalid adduct format 
#> 2024-08-12 17:32:49.399269 Returning 0 instead 
#> 2024-08-12 17:32:49.399394 Invalid adduct format 
#> 2024-08-12 17:32:49.399459 Returning 0 instead 
#> 2024-08-12 17:32:49.399601 Invalid adduct format 
#> 2024-08-12 17:32:49.399675 Returning 0 instead 
#> 2024-08-12 17:32:49.399801 Invalid adduct format 
#> 2024-08-12 17:32:49.39987 Returning 0 instead 
#> 2024-08-12 17:32:49.399999 Invalid adduct format 
#> 2024-08-12 17:32:49.400066 Returning 0 instead 
#> 2024-08-12 17:32:49.400189 Invalid adduct format 
#> 2024-08-12 17:32:49.400254 Returning 0 instead 
#> 2024-08-12 17:32:49.400378 Invalid adduct format 
#> 2024-08-12 17:32:49.400445 Returning 0 instead 
#> 2024-08-12 17:32:49.400569 Invalid adduct format 
#> 2024-08-12 17:32:49.400634 Returning 0 instead 
#> 2024-08-12 17:32:49.400757 Invalid adduct format 
#> 2024-08-12 17:32:49.400821 Returning 0 instead 
#> 2024-08-12 17:32:49.400948 Invalid adduct format 
#> 2024-08-12 17:32:49.401013 Returning 0 instead 
#> 2024-08-12 17:32:49.40114 Invalid adduct format 
#> 2024-08-12 17:32:49.401206 Returning 0 instead 
#> 2024-08-12 17:32:49.401331 Invalid adduct format 
#> 2024-08-12 17:32:49.401399 Returning 0 instead 
#> 2024-08-12 17:32:49.401523 Invalid adduct format 
#> 2024-08-12 17:32:49.401588 Returning 0 instead 
#> 2024-08-12 17:32:49.401711 Invalid adduct format 
#> 2024-08-12 17:32:49.401794 Returning 0 instead 
#> 2024-08-12 17:32:49.401949 Invalid adduct format 
#> 2024-08-12 17:32:49.402013 Returning 0 instead 
#> 2024-08-12 17:32:49.402142 Invalid adduct format 
#> 2024-08-12 17:32:49.402211 Returning 0 instead 
#> 2024-08-12 17:32:49.402347 Invalid adduct format 
#> 2024-08-12 17:32:49.402413 Returning 0 instead 
#> 2024-08-12 17:32:49.402536 Invalid adduct format 
#> 2024-08-12 17:32:49.402601 Returning 0 instead 
#> 2024-08-12 17:32:49.402721 Invalid adduct format 
#> 2024-08-12 17:32:49.402785 Returning 0 instead 
#> 2024-08-12 17:32:49.402906 Invalid adduct format 
#> 2024-08-12 17:32:49.40297 Returning 0 instead 
#> 2024-08-12 17:32:49.403093 Invalid adduct format 
#> 2024-08-12 17:32:49.403158 Returning 0 instead 
#> 2024-08-12 17:32:49.403281 Invalid adduct format 
#> 2024-08-12 17:32:49.403345 Returning 0 instead 
#> 2024-08-12 17:32:49.403467 Invalid adduct format 
#> 2024-08-12 17:32:49.403531 Returning 0 instead 
#> 2024-08-12 17:32:49.403677 Invalid adduct format 
#> 2024-08-12 17:32:49.403753 Returning 0 instead 
#> 2024-08-12 17:32:49.403873 Invalid adduct format 
#> 2024-08-12 17:32:49.403935 Returning 0 instead 
#> 2024-08-12 17:32:49.404065 Invalid adduct format 
#> 2024-08-12 17:32:49.404127 Returning 0 instead 
#> 2024-08-12 17:32:49.404249 Invalid adduct format 
#> 2024-08-12 17:32:49.404313 Returning 0 instead 
#> 2024-08-12 17:32:49.404437 Invalid adduct format 
#> 2024-08-12 17:32:49.404504 Returning 0 instead 
#> 2024-08-12 17:32:49.404629 Invalid adduct format 
#> 2024-08-12 17:32:49.404695 Returning 0 instead 
#> 2024-08-12 17:32:49.404852 Invalid adduct format 
#> 2024-08-12 17:32:49.404918 Returning 0 instead 
#> 2024-08-12 17:32:49.405042 Invalid adduct format 
#> 2024-08-12 17:32:49.405105 Returning 0 instead 
#> 2024-08-12 17:32:49.405231 Invalid adduct format 
#> 2024-08-12 17:32:49.405294 Returning 0 instead 
#> 2024-08-12 17:32:49.405416 Invalid adduct format 
#> 2024-08-12 17:32:49.405482 Returning 0 instead 
#> 2024-08-12 17:32:49.405604 Invalid adduct format 
#> 2024-08-12 17:32:49.405669 Returning 0 instead 
#> 2024-08-12 17:32:49.405817 Invalid adduct format 
#> 2024-08-12 17:32:49.405892 Returning 0 instead 
#> 2024-08-12 17:32:49.406024 Invalid adduct format 
#> 2024-08-12 17:32:49.406089 Returning 0 instead 
#> 2024-08-12 17:32:49.406213 Invalid adduct format 
#> 2024-08-12 17:32:49.406276 Returning 0 instead 
#> 2024-08-12 17:32:49.406397 Invalid adduct format 
#> 2024-08-12 17:32:49.406462 Returning 0 instead 
#> 2024-08-12 17:32:49.406587 Invalid adduct format 
#> 2024-08-12 17:32:49.406649 Returning 0 instead 
#> 2024-08-12 17:32:49.406774 Invalid adduct format 
#> 2024-08-12 17:32:49.406837 Returning 0 instead 
#> 2024-08-12 17:32:49.406959 Invalid adduct format 
#> 2024-08-12 17:32:49.407023 Returning 0 instead 
#> 2024-08-12 17:32:49.407145 Invalid adduct format 
#> 2024-08-12 17:32:49.407208 Returning 0 instead 
#> 2024-08-12 17:32:49.40733 Invalid adduct format 
#> 2024-08-12 17:32:49.407391 Returning 0 instead 
#> 2024-08-12 17:32:49.407511 Invalid adduct format 
#> 2024-08-12 17:32:49.40759 Returning 0 instead 
#> 2024-08-12 17:32:49.407724 Invalid adduct format 
#> 2024-08-12 17:32:49.407789 Returning 0 instead 
#> 2024-08-12 17:32:49.40791 Invalid adduct format 
#> 2024-08-12 17:32:49.407974 Returning 0 instead 
#> 2024-08-12 17:32:49.408097 Invalid adduct format 
#> 2024-08-12 17:32:49.408161 Returning 0 instead 
#> 2024-08-12 17:32:49.408283 Invalid adduct format 
#> 2024-08-12 17:32:49.408343 Returning 0 instead 
#> 2024-08-12 17:32:49.408466 Invalid adduct format 
#> 2024-08-12 17:32:49.408527 Returning 0 instead 
#> 2024-08-12 17:32:49.408648 Invalid adduct format 
#> 2024-08-12 17:32:49.40871 Returning 0 instead 
#> 2024-08-12 17:32:49.40883 Invalid adduct format 
#> 2024-08-12 17:32:49.408895 Returning 0 instead 
#> 2024-08-12 17:32:49.40902 Invalid adduct format 
#> 2024-08-12 17:32:49.409083 Returning 0 instead 
#> 2024-08-12 17:32:49.409205 Invalid adduct format 
#> 2024-08-12 17:32:49.409269 Returning 0 instead 
#> 2024-08-12 17:32:49.409391 Invalid adduct format 
#> 2024-08-12 17:32:49.409452 Returning 0 instead 
#> 2024-08-12 17:32:49.413662 joining within given rt tolerance 
#>  
#> 2024-08-12 17:32:49.41602 joining within given mz tol and filtering possible adducts 
#>  
#> 2024-08-12 17:32:49.428233 joining single adducts, in source dimers, and multicharged 
#>  
#> 2024-08-12 17:32:49.433357 adding chemical classification 
#> 2024-08-12 17:32:49.445364 MS1 annotation led to 
#>  0 annotations , on 
#>  0 features 
#> Warning: cannot open file '/home/runner/work/_temp/Library/timaparams/default/annotate_masses.yaml': No such file or directory
#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection