Annotate masses

This function annotates a feature table based on exact mass match. It requires a structural library, its metadata, and lists of adducts, clusters, and neutral losses to be considered. The polarity has to be pos or neg and retention time and mass tolerances should be given. The feature table is expected to be pre-formatted.

Usage

annotate_masses(
  features = get_params(step = "annotate_masses")$files$features$prepared,
  output_annotations = get_params(step =
    "annotate_masses")$files$annotations$prepared$structural$ms1,
  output_edges = get_params(step = "annotate_masses")$files$networks$spectral$edges$raw,
  name_source = get_params(step = "annotate_masses")$names$source,
  name_target = get_params(step = "annotate_masses")$names$target,
  library = get_params(step = "annotate_masses")$files$libraries$sop$merged$keys,
  str_stereo = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$stereo,
  str_met = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$metadata,
  str_nam = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$names,
  str_tax_cla = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$cla,
  str_tax_npc = get_params(step =
    "annotate_masses")$files$libraries$sop$merged$structures$taxonomies$npc,
  adducts_list = get_params(step = "annotate_masses")$ms$adducts,
  clusters_list = get_params(step = "annotate_masses")$ms$clusters,
  neutral_losses_list = get_params(step = "annotate_masses")$ms$neutral_losses,
  ms_mode = get_params(step = "annotate_masses")$ms$polarity,
  tolerance_ppm = get_params(step = "annotate_masses")$ms$tolerances$mass$ppm$ms1,
  tolerance_rt = get_params(step = "annotate_masses")$ms$tolerances$rt$adducts
)

Arguments

features

Table containing your previous annotation to complement

output_annotations

Output for mass based structural annotations

output_edges

Output for mass based edges

name_source

Name of the source features column

name_target

Name of the target features column

library

Library containing the keys

str_stereo

File containing structures stereo

str_met

File containing structures metadata

str_nam

File containing structures names

str_tax_cla

File containing Classyfire taxonomy

str_tax_npc

File containing NPClassifier taxonomy

adducts_list

List of adducts to be used

clusters_list

List of clusters to be used

neutral_losses_list

List of neutral losses to be used

ms_mode

Ionization mode. Must be 'pos' or 'neg'

tolerance_ppm

Tolerance to perform annotation. Should be <= 10 ppm

tolerance_rt

Tolerance to group adducts. Should be <= 0.05 minutes

Value

The path to the files containing MS1 annotations and edges

Examples

github <- "https://raw.githubusercontent.com/"
repo <- "taxonomicallyinformedannotation/tima-example-files/"
data_interim <- "main/data/interim/"
dir <- paste0(github, repo, data_interim)
annotate_masses(
  features = paste0(dir, "features/example_features.tsv"),
  library = paste0(dir, "libraries/sop/merged/keys.tsv"),
  str_stereo = paste0(dir, "libraries/sop/merged/structures/stereo.tsv"),
  str_met = paste0(dir, "libraries/sop/merged/structures/metadata.tsv"),
  str_nam = paste0(dir, "libraries/sop/merged/structures/names.tsv"),
  str_tax_cla = paste0(dir, "libraries/sop/merged/structures/taxonomies/classyfire.tsv"),
  str_tax_npc = paste0(dir, "libraries/sop/merged/structures/taxonomies/npc.tsv"),
  adducts_list = list(pos = c("[M+H]+"), neg = c("[M-H]-")),
  clusters_list = list(pos = c("H2O"), neg = c("H2O")),
  neutral_losses_list = c("C6H10O4 (methylpentose/desoxyhexose-H2O)"),
  ms_mode = "pos",
  tolerance_ppm = 10,
  tolerance_rt = 0.01,
  output_annotations = "output_annotations.tsv",
  output_edges = "output_edges.tsv"
)
#> 2024-08-12 15:01:15.825063 filtering desired adducts and adding mz tolerance 
#>  
#> 2024-08-12 15:01:15.835096 No previously attributed adducts detected 
#> 2024-08-12 15:01:15.842166 calculating rt tolerance ... 
#>  
#> 2024-08-12 15:01:15.843772 joining within given rt tolerance 
#>  
#> 2024-08-12 15:01:15.852457 adding delta mz tolerance for single charge adducts 
#>  
#> 2024-08-12 15:01:15.857083 forming adducts and clusters 
#>  
#> 2024-08-12 15:01:15.92025 calculating delta mz for single charge adducts and clusters 
#>  
#> 2024-08-12 15:01:15.934274 joining within given delta mz tolerance (neutral losses) 
#>  
#> 2024-08-12 15:01:15.939178 joining within given delta mz tolerance (adducts) 
#>  
#> 2024-08-12 15:01:15.9492 keeping initial and destination feature 
#>  
#> 2024-08-12 15:01:15.955964 joining with initial results (adducts) 
#>  
#> 2024-08-12 15:01:15.958848 joining with initial results (neutral losses) 
#>  
#> 2024-08-12 15:01:16.004198 joining within given mz tol to exact mass library 
#>  
#> 2024-08-12 15:01:16.009807 keeping unique exact masses and molecular formulas 
#>  
#> 2024-08-12 15:01:16.018944 joining exact masses with single charge adducts 
#>  
#> 2024-08-12 15:01:16.019033 Geting back to M 
#>  
#> 2024-08-12 15:01:16.022463 calculating multicharged and in source dimers 
#>  
#> 2024-08-12 15:01:16.0309 Invalid adduct format 
#> 2024-08-12 15:01:16.030999 Returning 0 instead 
#> 2024-08-12 15:01:16.03115 Invalid adduct format 
#> 2024-08-12 15:01:16.03122 Returning 0 instead 
#> 2024-08-12 15:01:16.031356 Invalid adduct format 
#> 2024-08-12 15:01:16.03142 Returning 0 instead 
#> 2024-08-12 15:01:16.03158 Invalid adduct format 
#> 2024-08-12 15:01:16.031647 Returning 0 instead 
#> 2024-08-12 15:01:16.031776 Invalid adduct format 
#> 2024-08-12 15:01:16.031839 Returning 0 instead 
#> 2024-08-12 15:01:16.031967 Invalid adduct format 
#> 2024-08-12 15:01:16.032031 Returning 0 instead 
#> 2024-08-12 15:01:16.032159 Invalid adduct format 
#> 2024-08-12 15:01:16.032221 Returning 0 instead 
#> 2024-08-12 15:01:16.032346 Invalid adduct format 
#> 2024-08-12 15:01:16.032409 Returning 0 instead 
#> 2024-08-12 15:01:16.032535 Invalid adduct format 
#> 2024-08-12 15:01:16.032598 Returning 0 instead 
#> 2024-08-12 15:01:16.032718 Invalid adduct format 
#> 2024-08-12 15:01:16.03278 Returning 0 instead 
#> 2024-08-12 15:01:16.032905 Invalid adduct format 
#> 2024-08-12 15:01:16.03297 Returning 0 instead 
#> 2024-08-12 15:01:16.033094 Invalid adduct format 
#> 2024-08-12 15:01:16.033157 Returning 0 instead 
#> 2024-08-12 15:01:16.03328 Invalid adduct format 
#> 2024-08-12 15:01:16.033342 Returning 0 instead 
#> 2024-08-12 15:01:16.033467 Invalid adduct format 
#> 2024-08-12 15:01:16.03353 Returning 0 instead 
#> 2024-08-12 15:01:16.033652 Invalid adduct format 
#> 2024-08-12 15:01:16.033714 Returning 0 instead 
#> 2024-08-12 15:01:16.033835 Invalid adduct format 
#> 2024-08-12 15:01:16.033898 Returning 0 instead 
#> 2024-08-12 15:01:16.034022 Invalid adduct format 
#> 2024-08-12 15:01:16.034085 Returning 0 instead 
#> 2024-08-12 15:01:16.034206 Invalid adduct format 
#> 2024-08-12 15:01:16.034271 Returning 0 instead 
#> 2024-08-12 15:01:16.034395 Invalid adduct format 
#> 2024-08-12 15:01:16.034458 Returning 0 instead 
#> 2024-08-12 15:01:16.03458 Invalid adduct format 
#> 2024-08-12 15:01:16.03464 Returning 0 instead 
#> 2024-08-12 15:01:16.034759 Invalid adduct format 
#> 2024-08-12 15:01:16.034822 Returning 0 instead 
#> 2024-08-12 15:01:16.034946 Invalid adduct format 
#> 2024-08-12 15:01:16.035015 Returning 0 instead 
#> 2024-08-12 15:01:16.035138 Invalid adduct format 
#> 2024-08-12 15:01:16.035202 Returning 0 instead 
#> 2024-08-12 15:01:16.035322 Invalid adduct format 
#> 2024-08-12 15:01:16.035383 Returning 0 instead 
#> 2024-08-12 15:01:16.035523 Invalid adduct format 
#> 2024-08-12 15:01:16.03559 Returning 0 instead 
#> 2024-08-12 15:01:16.035714 Invalid adduct format 
#> 2024-08-12 15:01:16.035779 Returning 0 instead 
#> 2024-08-12 15:01:16.035905 Invalid adduct format 
#> 2024-08-12 15:01:16.035973 Returning 0 instead 
#> 2024-08-12 15:01:16.036099 Invalid adduct format 
#> 2024-08-12 15:01:16.036162 Returning 0 instead 
#> 2024-08-12 15:01:16.036287 Invalid adduct format 
#> 2024-08-12 15:01:16.036351 Returning 0 instead 
#> 2024-08-12 15:01:16.036472 Invalid adduct format 
#> 2024-08-12 15:01:16.036537 Returning 0 instead 
#> 2024-08-12 15:01:16.036659 Invalid adduct format 
#> 2024-08-12 15:01:16.036722 Returning 0 instead 
#> 2024-08-12 15:01:16.03685 Invalid adduct format 
#> 2024-08-12 15:01:16.036919 Returning 0 instead 
#> 2024-08-12 15:01:16.037046 Invalid adduct format 
#> 2024-08-12 15:01:16.037113 Returning 0 instead 
#> 2024-08-12 15:01:16.037241 Invalid adduct format 
#> 2024-08-12 15:01:16.037308 Returning 0 instead 
#> 2024-08-12 15:01:16.037434 Invalid adduct format 
#> 2024-08-12 15:01:16.037498 Returning 0 instead 
#> 2024-08-12 15:01:16.037618 Invalid adduct format 
#> 2024-08-12 15:01:16.037685 Returning 0 instead 
#> 2024-08-12 15:01:16.037809 Invalid adduct format 
#> 2024-08-12 15:01:16.037874 Returning 0 instead 
#> 2024-08-12 15:01:16.037994 Invalid adduct format 
#> 2024-08-12 15:01:16.038055 Returning 0 instead 
#> 2024-08-12 15:01:16.038176 Invalid adduct format 
#> 2024-08-12 15:01:16.03824 Returning 0 instead 
#> 2024-08-12 15:01:16.038361 Invalid adduct format 
#> 2024-08-12 15:01:16.038422 Returning 0 instead 
#> 2024-08-12 15:01:16.038542 Invalid adduct format 
#> 2024-08-12 15:01:16.038602 Returning 0 instead 
#> 2024-08-12 15:01:16.038721 Invalid adduct format 
#> 2024-08-12 15:01:16.038787 Returning 0 instead 
#> 2024-08-12 15:01:16.038918 Invalid adduct format 
#> 2024-08-12 15:01:16.038983 Returning 0 instead 
#> 2024-08-12 15:01:16.039107 Invalid adduct format 
#> 2024-08-12 15:01:16.039172 Returning 0 instead 
#> 2024-08-12 15:01:16.039293 Invalid adduct format 
#> 2024-08-12 15:01:16.039358 Returning 0 instead 
#> 2024-08-12 15:01:16.039502 Invalid adduct format 
#> 2024-08-12 15:01:16.039575 Returning 0 instead 
#> 2024-08-12 15:01:16.039701 Invalid adduct format 
#> 2024-08-12 15:01:16.039767 Returning 0 instead 
#> 2024-08-12 15:01:16.039893 Invalid adduct format 
#> 2024-08-12 15:01:16.039957 Returning 0 instead 
#> 2024-08-12 15:01:16.040079 Invalid adduct format 
#> 2024-08-12 15:01:16.040144 Returning 0 instead 
#> 2024-08-12 15:01:16.040266 Invalid adduct format 
#> 2024-08-12 15:01:16.040331 Returning 0 instead 
#> 2024-08-12 15:01:16.040456 Invalid adduct format 
#> 2024-08-12 15:01:16.040523 Returning 0 instead 
#> 2024-08-12 15:01:16.040649 Invalid adduct format 
#> 2024-08-12 15:01:16.040714 Returning 0 instead 
#> 2024-08-12 15:01:16.040839 Invalid adduct format 
#> 2024-08-12 15:01:16.040901 Returning 0 instead 
#> 2024-08-12 15:01:16.041025 Invalid adduct format 
#> 2024-08-12 15:01:16.041089 Returning 0 instead 
#> 2024-08-12 15:01:16.041211 Invalid adduct format 
#> 2024-08-12 15:01:16.041275 Returning 0 instead 
#> 2024-08-12 15:01:16.041396 Invalid adduct format 
#> 2024-08-12 15:01:16.041461 Returning 0 instead 
#> 2024-08-12 15:01:16.041582 Invalid adduct format 
#> 2024-08-12 15:01:16.041646 Returning 0 instead 
#> 2024-08-12 15:01:16.041771 Invalid adduct format 
#> 2024-08-12 15:01:16.041837 Returning 0 instead 
#> 2024-08-12 15:01:16.041962 Invalid adduct format 
#> 2024-08-12 15:01:16.04203 Returning 0 instead 
#> 2024-08-12 15:01:16.042154 Invalid adduct format 
#> 2024-08-12 15:01:16.042218 Returning 0 instead 
#> 2024-08-12 15:01:16.042339 Invalid adduct format 
#> 2024-08-12 15:01:16.042403 Returning 0 instead 
#> 2024-08-12 15:01:16.042527 Invalid adduct format 
#> 2024-08-12 15:01:16.042589 Returning 0 instead 
#> 2024-08-12 15:01:16.042719 Invalid adduct format 
#> 2024-08-12 15:01:16.042786 Returning 0 instead 
#> 2024-08-12 15:01:16.042912 Invalid adduct format 
#> 2024-08-12 15:01:16.042977 Returning 0 instead 
#> 2024-08-12 15:01:16.043099 Invalid adduct format 
#> 2024-08-12 15:01:16.043162 Returning 0 instead 
#> 2024-08-12 15:01:16.043283 Invalid adduct format 
#> 2024-08-12 15:01:16.043346 Returning 0 instead 
#> 2024-08-12 15:01:16.043502 Invalid adduct format 
#> 2024-08-12 15:01:16.043579 Returning 0 instead 
#> 2024-08-12 15:01:16.043711 Invalid adduct format 
#> 2024-08-12 15:01:16.043777 Returning 0 instead 
#> 2024-08-12 15:01:16.043904 Invalid adduct format 
#> 2024-08-12 15:01:16.04397 Returning 0 instead 
#> 2024-08-12 15:01:16.044094 Invalid adduct format 
#> 2024-08-12 15:01:16.044159 Returning 0 instead 
#> 2024-08-12 15:01:16.044284 Invalid adduct format 
#> 2024-08-12 15:01:16.044348 Returning 0 instead 
#> 2024-08-12 15:01:16.044471 Invalid adduct format 
#> 2024-08-12 15:01:16.044533 Returning 0 instead 
#> 2024-08-12 15:01:16.044655 Invalid adduct format 
#> 2024-08-12 15:01:16.044728 Returning 0 instead 
#> 2024-08-12 15:01:16.044851 Invalid adduct format 
#> 2024-08-12 15:01:16.044913 Returning 0 instead 
#> 2024-08-12 15:01:16.045036 Invalid adduct format 
#> 2024-08-12 15:01:16.045099 Returning 0 instead 
#> 2024-08-12 15:01:16.045221 Invalid adduct format 
#> 2024-08-12 15:01:16.045287 Returning 0 instead 
#> 2024-08-12 15:01:16.045409 Invalid adduct format 
#> 2024-08-12 15:01:16.045471 Returning 0 instead 
#> 2024-08-12 15:01:16.045596 Invalid adduct format 
#> 2024-08-12 15:01:16.045699 Returning 0 instead 
#> 2024-08-12 15:01:16.045859 Invalid adduct format 
#> 2024-08-12 15:01:16.045923 Returning 0 instead 
#> 2024-08-12 15:01:16.046048 Invalid adduct format 
#> 2024-08-12 15:01:16.046111 Returning 0 instead 
#> 2024-08-12 15:01:16.046237 Invalid adduct format 
#> 2024-08-12 15:01:16.0463 Returning 0 instead 
#> 2024-08-12 15:01:16.046424 Invalid adduct format 
#> 2024-08-12 15:01:16.046489 Returning 0 instead 
#> 2024-08-12 15:01:16.046611 Invalid adduct format 
#> 2024-08-12 15:01:16.046675 Returning 0 instead 
#> 2024-08-12 15:01:16.046798 Invalid adduct format 
#> 2024-08-12 15:01:16.04686 Returning 0 instead 
#> 2024-08-12 15:01:16.046982 Invalid adduct format 
#> 2024-08-12 15:01:16.047043 Returning 0 instead 
#> 2024-08-12 15:01:16.047163 Invalid adduct format 
#> 2024-08-12 15:01:16.047227 Returning 0 instead 
#> 2024-08-12 15:01:16.047349 Invalid adduct format 
#> 2024-08-12 15:01:16.047412 Returning 0 instead 
#> 2024-08-12 15:01:16.04756 Invalid adduct format 
#> 2024-08-12 15:01:16.047627 Returning 0 instead 
#> 2024-08-12 15:01:16.047751 Invalid adduct format 
#> 2024-08-12 15:01:16.047814 Returning 0 instead 
#> 2024-08-12 15:01:16.047934 Invalid adduct format 
#> 2024-08-12 15:01:16.047997 Returning 0 instead 
#> 2024-08-12 15:01:16.048116 Invalid adduct format 
#> 2024-08-12 15:01:16.048179 Returning 0 instead 
#> 2024-08-12 15:01:16.0483 Invalid adduct format 
#> 2024-08-12 15:01:16.048362 Returning 0 instead 
#> 2024-08-12 15:01:16.048484 Invalid adduct format 
#> 2024-08-12 15:01:16.048549 Returning 0 instead 
#> 2024-08-12 15:01:16.04867 Invalid adduct format 
#> 2024-08-12 15:01:16.048735 Returning 0 instead 
#> 2024-08-12 15:01:16.048856 Invalid adduct format 
#> 2024-08-12 15:01:16.048924 Returning 0 instead 
#> 2024-08-12 15:01:16.049047 Invalid adduct format 
#> 2024-08-12 15:01:16.04911 Returning 0 instead 
#> 2024-08-12 15:01:16.049232 Invalid adduct format 
#> 2024-08-12 15:01:16.049294 Returning 0 instead 
#> 2024-08-12 15:01:16.04942 Invalid adduct format 
#> 2024-08-12 15:01:16.049486 Returning 0 instead 
#> 2024-08-12 15:01:16.049608 Invalid adduct format 
#> 2024-08-12 15:01:16.049671 Returning 0 instead 
#> 2024-08-12 15:01:16.049795 Invalid adduct format 
#> 2024-08-12 15:01:16.049857 Returning 0 instead 
#> 2024-08-12 15:01:16.049979 Invalid adduct format 
#> 2024-08-12 15:01:16.050041 Returning 0 instead 
#> 2024-08-12 15:01:16.05383 joining within given rt tolerance 
#>  
#> 2024-08-12 15:01:16.056041 joining within given mz tol and filtering possible adducts 
#>  
#> 2024-08-12 15:01:16.06755 joining single adducts, in source dimers, and multicharged 
#>  
#> 2024-08-12 15:01:16.072399 adding chemical classification 
#> 2024-08-12 15:01:16.083976 MS1 annotation led to 
#>  0 annotations , on 
#>  0 features 
#> Warning: cannot open file '/home/runner/work/_temp/Library/timaparams/default/annotate_masses.yaml': No such file or directory
#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection